+++ amipy/amiwrapper.py
@@ -17,6 +17,7 @@ from typing import Iterable, Tuple
 import numpy as np
 
 from amipy.ami import Ami
+from amipy.utils import SuppressStream
 
 
 @unique
@@ -449,18 +450,3 @@ def execute_function(function, arguments, detail=""):
                 msg = f"MODFLOW 6 BMI, exception in: {function.__name__} ({detail})"
                 raise Exception(msg)
 
-
-class SuppressStream:
-    def __init__(self, stream=sys.stderr):
-        self.orig_stream_fileno = stream.fileno()
-
-    def __enter__(self):
-        self.orig_stream_dup = os.dup(self.orig_stream_fileno)
-        self.devnull = open(os.devnull, "w")
-        os.dup2(self.devnull.fileno(), self.orig_stream_fileno)
-
-    def __exit__(self, type, value, traceback):
-        os.close(self.orig_stream_fileno)
-        os.dup2(self.orig_stream_dup, self.orig_stream_fileno)
-        os.close(self.orig_stream_dup)
-        self.devnull.close()
