+++ amipy/utils.py
@@ -1,0 +1,19 @@
+import sys
+import os
+
+
+# copied from https://stackoverflow.com/a/57677370
+class SuppressStream:
+    def __init__(self, stream=sys.stderr):
+        self.orig_stream_fileno = stream.fileno()
+
+    def __enter__(self):
+        self.orig_stream_dup = os.dup(self.orig_stream_fileno)
+        self.devnull = open(os.devnull, "w")
+        os.dup2(self.devnull.fileno(), self.orig_stream_fileno)
+
+    def __exit__(self, type, value, traceback):
+        os.close(self.orig_stream_fileno)
+        os.dup2(self.orig_stream_dup, self.orig_stream_fileno)
+        os.close(self.orig_stream_dup)
+        self.devnull.close()
